requirements.txt
discord.py==2.0.1
# git+https://github.com/fluxblox/discord.py
aiodns==3.0.0
cchardet==2.1.7
psutil==5.9.0
PyJWT==2.4.0
python-dateutil==2.8.2
jsonpath_ng==1.5.3
aiotrello==0.0.7.6
aredis==1.1.8
hiredis==2.0.0
#sentry-sdk==0.18.0
python-benedict==0.25.0
dnspython==2.0.0
motor==2.5.1
async_timeout==4.0.2